+++ pygeos/coordinates.py
@@ -20,8 +20,8 @@ def apply(geometry, transformation, include_z=False):
         A function that transforms a (N, 2) or (N, 3) ndarray of float64 to
         another (N, 2) or (N, 3) ndarray of float64.
     include_z : bool, default False
-        Whether to include the third dimension in the coordinates array
-        that is passed to the `transformation` function. If True, and a
+        If True, include the third dimension in the coordinates array
+        that is passed to the `transformation` function. If a
         geometry has no third dimension, the z-coordinates passed to the
         function will be NaN.
 
@@ -44,7 +44,7 @@ def apply(geometry, transformation, include_z=False):
     <pygeos.Geometry POINT Z (1 1 1)>
     """
     geometry_arr = np.array(geometry, dtype=np.object_)  # makes a copy
-    coordinates = lib.get_coordinates(geometry_arr, include_z)
+    coordinates = lib.get_coordinates(geometry_arr, include_z, False)
     new_coordinates = transformation(coordinates)
     # check the array to yield understandable error messages
     if not isinstance(new_coordinates, np.ndarray):
@@ -87,7 +87,7 @@ def count_coordinates(geometry):
     return lib.count_coordinates(np.asarray(geometry, dtype=np.object_))
 
 
-def get_coordinates(geometry, include_z=False):
+def get_coordinates(geometry, include_z=False, return_index=False):
     """Gets coordinates from a geometry array as an array of floats.
 
     The shape of the returned array is (N, 2), with N being the number of
@@ -99,8 +99,12 @@ def get_coordinates(geometry, include_z=False):
     ----------
     geometry : Geometry or array_like
     include_z : bool, default False
-        Whether to include the third dimension in the output. If True, and a
-        geometry has no third dimension, the z-coordinates will be NaN.
+        If, True include the third dimension in the output. If a geometry
+        has no third dimension, the z-coordinates will be NaN.
+    return_index : bool, default False
+        If True, also return the index of each returned geometry as a separate
+        ndarray of integers. For multidimensional arrays, this indexes into the
+        flattened array (in C contiguous order).
 
     Examples
     --------
@@ -117,8 +121,16 @@ def get_coordinates(geometry, include_z=False):
     [[0.0, 0.0]]
     >>> get_coordinates(Geometry("POINT Z (0 0 0)"), include_z=True).tolist()
     [[0.0, 0.0, 0.0]]
+
+    When return_index=True, indexes are returned also:
+    >>> geometries = [Geometry("LINESTRING (2 2, 4 4)"), Geometry("POINT (0 0)")]
+    >>> coordinates, index = get_coordinates(geometries, return_index=True)
+    >>> coordinates.tolist(), index.tolist()
+    ([[2.0, 2.0], [4.0, 4.0], [0.0, 0.0]], [0, 0, 1])
     """
-    return lib.get_coordinates(np.asarray(geometry, dtype=np.object_), include_z)
+    return lib.get_coordinates(
+        np.asarray(geometry, dtype=np.object_), include_z, return_index
+    )
 
 
 def set_coordinates(geometry, coordinates):
@@ -159,9 +171,8 @@ def set_coordinates(geometry, coordinates):
             "The coordinate array should have dimension of 2 "
             "(has {})".format(coordinates.ndim)
         )
-    if (
-        (coordinates.shape[0] != lib.count_coordinates(geometry_arr))
-        or (coordinates.shape[1] not in {2, 3})
+    if (coordinates.shape[0] != lib.count_coordinates(geometry_arr)) or (
+        coordinates.shape[1] not in {2, 3}
     ):
         raise ValueError(
             "The coordinate array has an invalid shape {}".format(coordinates.shape)